[ALPS05710742] reviser: support rv
1. support rv
2. support rpmsg

MTK-Commit-Id: c22f51195fb0ccf4f9fbd7abc1a0e3fe222dc9f3

CR-Id: ALPS05710742
Feature: [Module] APUSYS

+++ drivers/misc/mediatek/apusys/apusys_rv/apu_config.h
@@ -38,7 +38,7 @@ struct logger_init_info {
 
 struct reviser_init_info {
 	uint32_t boundary;
-	uint32_t dram[32];
+	uint64_t dram[32];
 } __attribute__ ((__packed__));
 
 struct mvpu_preempt_data {
